Refactored all Rule related classes to their own namespace
