+++ PasswordRulesSharp.DemoTool/Program.cs
@@ -1,6 +1,8 @@
-﻿const string rawRule = "minlength: 8; maxlength: 10; required: lower; required: upper; required: digit; required: [-]";
+﻿using PasswordRulesSharp.Rules;
 
-var rule = new PasswordRulesSharp.Parser.StringRule(rawRule);
+const string rawRule = "minlength: 8; maxlength: 10; required: lower; required: upper; required: digit; required: [-]";
+
+var rule = new StringRule(rawRule);
 
 const string samplePassword1 = "Hello2AndSomeMoreText-";
