+++ PasswordRulesSharp.Tests/Generator/ChooseLengthTests.cs
@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using PasswordRulesSharp.Rules;
 
 namespace PasswordRulesSharp.Tests.Generator
 {
@@ -13,7 +14,7 @@ namespace PasswordRulesSharp.Tests.Generator
         [TestCase("required: upper", 20)]
         public void ChooseLength(string rule, int expectedLength)
         {
-            var parsedRule = new PasswordRulesSharp.Parser.StringRule(rule);
+            var parsedRule = new StringRule(rule);
 
             var generator = new PasswordRulesSharp.Generator.Generator(parsedRule);
