+++ PasswordRulesSharp.Tests/Generator/GeneratedPasswordTests.cs
@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using PasswordRulesSharp.Rules;
 
 namespace PasswordRulesSharp.Tests.Generator
 {
@@ -13,7 +14,7 @@ namespace PasswordRulesSharp.Tests.Generator
         [TestCase("required: upper")]
         public void LengthMatchesExpected(string rule)
         {
-            var parsedRule = new PasswordRulesSharp.Parser.StringRule(rule);
+            var parsedRule = new StringRule(rule);
 
             var generator = new PasswordRulesSharp.Generator.Generator(parsedRule);
 
@@ -27,7 +28,7 @@ namespace PasswordRulesSharp.Tests.Generator
         [TestCase("required: lower")]
         public void RequiredLowerContainsIt(string rule)
         {
-            var parsedRule = new PasswordRulesSharp.Parser.StringRule(rule);
+            var parsedRule = new StringRule(rule);
             var generator = new PasswordRulesSharp.Generator.Generator(parsedRule);
             var password = generator.GeneratePassword();
 
@@ -54,7 +55,7 @@ namespace PasswordRulesSharp.Tests.Generator
         [TestCase("required: upper")]
         public void RequiredUpperContainsIt(string rule)
         {
-            var parsedRule = new PasswordRulesSharp.Parser.StringRule(rule);
+            var parsedRule = new StringRule(rule);
             var generator = new PasswordRulesSharp.Generator.Generator(parsedRule);
             var password = generator.GeneratePassword();
 
@@ -81,7 +82,7 @@ namespace PasswordRulesSharp.Tests.Generator
         [TestCase("required: digit")]
         public void RequiredDigitContainsIt(string rule)
         {
-            var parsedRule = new PasswordRulesSharp.Parser.StringRule(rule);
+            var parsedRule = new StringRule(rule);
             var generator = new PasswordRulesSharp.Generator.Generator(parsedRule);
             var password = generator.GeneratePassword();
 
@@ -109,7 +110,7 @@ namespace PasswordRulesSharp.Tests.Generator
         {
             var rule = $"required: [{chars}]";
 
-            var parsedRule = new PasswordRulesSharp.Parser.StringRule(rule);
+            var parsedRule = new StringRule(rule);
             var generator = new PasswordRulesSharp.Generator.Generator(parsedRule);
             var password = generator.GeneratePassword();
