+++ PasswordRulesSharp/Rules/Rule.cs
@@ -1,8 +1,9 @@
 using System.Collections.Generic;
 using NodaTime;
 using PasswordRulesSharp.Models;
+using PasswordRulesSharp.Parser;
 
-namespace PasswordRulesSharp.Parser;
+namespace PasswordRulesSharp.Rules;
 
 public class Rule : IRule
 {
